Move bootstrap code into its own module
