+++ validator/src/lib.rs
@@ -17,6 +17,7 @@ use {
 };
 
 pub mod admin_rpc_service;
+pub mod bootstrap;
 pub mod dashboard;
 
 #[cfg(unix)]
